requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
openpyxl==3.1.5
pdfkit==1.0.0
gunicorn==22.0.0